Update the discounts flow
Change the discounts service tests and DiscountController

Apply the OneHelpSDK to the proxy discounts controller and tests

+++ proxy/app/Http/Controllers/DiscountsController.php
@@ -7,11 +7,10 @@ use Exception;
 use App\User;
 use App\Discount;
 
-use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Http\JsonResponse;
-
+use OneHelpSDK\Facades\Discounts;
 class DiscountsController extends Controller
 {
     public function list(): JsonResponse
@@ -23,6 +22,10 @@ class DiscountsController extends Controller
     {
         try {
             $data = $request->all();
+            $discountApi = Discounts::create($data);
+            if (Discounts::getStatus() !== 201) {
+                throw new Exception('An error occurred when the discount was saved');
+            }
             $user = User::firstOrCreate(['email' => $data['userEmail']]);
             if (empty($user->id)) {
                 $user->save();
@@ -30,7 +33,7 @@ class DiscountsController extends Controller
             $data['user_id'] = $user->id;
             unset($data['userEmail']);
             $discount = new Discount($data);
-            $discount->reference = Str::uuid();
+            $discount->reference = $discountApi->_id;
             $discount->save();
             return response()->json(null, Response::HTTP_CREATED);
         } catch (Exception $e) {
@@ -42,7 +45,11 @@ class DiscountsController extends Controller
     {
         try {
             $data = $request->all();
-            $discount = Discount::firstOrFail($id);
+            $discount = Discount::findOrFail($id);
+            Discounts::update($discount->reference, $data);
+            if (Discounts::getStatus() !== 201) {
+                throw new Exception('An error occurred when the discount was updated');
+            }
             $this->fillWithData($discount, $data);
             $discount->save();
             return response()->json(null, Response::HTTP_NO_CONTENT);
@@ -71,7 +78,11 @@ class DiscountsController extends Controller
     public function delete(int $id): JsonResponse
     {
         try {
-            $discount = Discount::firstOrFail($id);
+            $discount = Discount::findOrFail($id);
+            Discounts::delete($discount->reference);
+            if (Discounts::getStatus() !== 204) {
+                throw new Exception('An error occurred when the request was updated');
+            }
             $discount->delete();
             return response()->json(null, Response::HTTP_NO_CONTENT);
         } catch (Exception $e) {
